test(pytest): fix conftest shadowing and scope testpaths so bare pytest works
Bare `pytest` loaded docs/conftest.py alongside tests/conftest.py, and the
test modules' `from conftest import HAS_HDF5 / requires_hdf5` resolved to the
wrong conftest, giving 28 ImportError collection errors.

Move the HDF5 detection logic into tests/hdf5_support.py and import it via the
package-qualified path `from tests.hdf5_support import ...` (the same pattern
tests/rvt_fixtures.py already uses). conftest.py re-exports the symbols for
backwards compatibility. Update every test module that did
`from conftest import ...` accordingly.

Narrow default testpaths to ["tests"] so bare `pytest` runs the test suite
only. docs (markdown-docs) and examples (nbmake) need extra deps plus data and
are invoked explicitly in CI, so drop --markdown-docs from default addopts;
docs can still be run with `pytest --markdown-docs docs/`.

Bare `pytest` now collects cleanly: 70 passed, 72 skipped, 0 errors.
