+++ tests/test_clean_output.py
@@ -7,7 +7,7 @@ import evlib
 import io
 from contextlib import redirect_stderr
 
-from conftest import requires_hdf5
+from tests.hdf5_support import requires_hdf5
 
 pytestmark = requires_hdf5
