+++ tests/test_hdf5_save_cross_platform.py
@@ -11,7 +11,7 @@ import os
 import numpy as np
 import pytest
 
-from conftest import requires_hdf5
+from tests.hdf5_support import requires_hdf5
 
 pytestmark = requires_hdf5
