+++ tests/test_load_events_validation.py
@@ -30,7 +30,7 @@ from typing import Dict, Any
 import polars as pl
 import pytest
 
-from conftest import requires_hdf5
+from tests.hdf5_support import requires_hdf5
 
 # Import validation helpers from tests directory
 try:
