+++ tests/test_representations.py
@@ -7,7 +7,7 @@ import evlib.representations as evr
 import tempfile
 import os
 
-from conftest import requires_hdf5
+from tests.hdf5_support import requires_hdf5
 
 pytestmark = requires_hdf5
