+++ tests/test_visualization.py
@@ -9,7 +9,7 @@ import numpy as np
 from pathlib import Path
 import logging
 
-from conftest import requires_hdf5
+from tests.hdf5_support import requires_hdf5
 
 pytestmark = requires_hdf5
